+++ src/app/models/Shared.ts
@@ -1,0 +1,4 @@
+export class Pair<S,T> {
+    first: S;
+    second: T;
+}