+++ src/app/modules/resourceMetadataModule.ts
@@ -3,6 +3,7 @@ import { NgModule } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 import { ImportPatternModal } from '../resourceMetadata/modals/importPatternModal';
 import { MetadataAssociationEditorModal } from '../resourceMetadata/modals/metadataAssociationEditorModal';
+import { MetadataFactoryPatternSelectionModal } from '../resourceMetadata/modals/metadataFactoryPatternSelectionModal';
 import { MetadataPatternEditorModal } from '../resourceMetadata/modals/metadataPatternEditorModal';
 import { MetadataPatternLibraryModal } from '../resourceMetadata/modals/metadataPatternLibraryModal';
 import { ResourceMetadataComponent } from '../resourceMetadata/resourceMetadataComponent';
@@ -15,8 +16,9 @@ import { SharedModule } from './sharedModule';
         ImportPatternModal,
         ResourceMetadataComponent,
         MetadataAssociationEditorModal,
+        MetadataFactoryPatternSelectionModal,
         MetadataPatternEditorModal,
-        MetadataPatternLibraryModal
+        MetadataPatternLibraryModal,
     ],
     exports: [
         ResourceMetadataComponent
@@ -24,6 +26,7 @@ import { SharedModule } from './sharedModule';
     entryComponents: [
         ImportPatternModal,
         MetadataAssociationEditorModal,
+        MetadataFactoryPatternSelectionModal,
         MetadataPatternEditorModal,
         MetadataPatternLibraryModal
     ]